+++ src/OrchardCore/OrchardCore.Infrastructure/OrchardCore.Infrastructure.csproj
@@ -9,6 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="MailKit" />
     <PackageReference Include="MessagePack" />
   </ItemGroup>
